build: source.js 생성 후 자잘한 변수와 함수 정리
