+++ js/source.js
@@ -1,0 +1,15 @@
+// 요소 변수 저장
+const button = document.querySelector(".btnSubmit");
+const input = document.querySelectorAll("input");
+const useremail = document.querySelector("#useremail");
+const username = document.querySelector("#username");
+const userpw = document.querySelector("#userpw");
+const pwcheck = document.querySelector("#pwcheck");
+
+// 이메일 형식 조건
+export function emailCheck(email) {    
+    let regex=/([\w-\.]+)@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.)|(([\w-]+\.)+))([a-zA-Z]{2,4}|[0-9]{1,3})(\]?)$/;
+    return (email != '' && email != 'undefined' && regex.test(email)); 
+};
+
+export { button, input, useremail, username, userpw, pwcheck };
